requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
google-generativeai==0.3.1
httpx==0.25.2
websockets==12.0
aiohttp==3.9.1
python-multipart==0.0.6
torch>=2.0.0
silero-vad>=4.0.0
numpy>=1.24.0
pydub>=0.25.1

